pyproject.toml
[project]
name = "podcast-insights"
version = "0.1.0"
description = "YouTube Podcast Insights Agent — Temporal + FastAPI + Gemini"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.19.0",
    "protobuf>=5.29.3,<6.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "youtube-transcript-api>=1.0.0",
    "boto3>=1.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "models", "workflows", "activities"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
